Get the tests passing again
