+++ lib/acidic_job/errors.rb
@@ -12,4 +12,5 @@ module AcidicJob
   class UnknownJobAdapter < Error; end
   class UnknownAwaitedJob < Error; end
   class TooManyParametersForStepMethod < Error; end
+  class UnserializableValue < Error; end
 end
