+++ test/combustion/app/models/notification.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "noticed"
+
 class Notification < ApplicationRecord
   include Noticed::Model
 end
