+++ test/combustion/config/database.yml
@@ -1,3 +1,6 @@
 test:
   adapter: sqlite3
   database: database.sqlite
+  flags: <%= SQLite3::Constants::Open::READWRITE |
+             SQLite3::Constants::Open::CREATE |
+             SQLite3::Constants::Open::SHAREDCACHE %>
